@@ -1,23 +1,23 @@
 {
   "name": "ghassen-rjab-portfolio",
   "description": "Ghassen Rjab's portfolio",
-  "version": "1.3.1",
+  "version": "1.4.0",
   "author": "Ghassen Rjab <rjab.ghassen@gmail.com>",
   "dependencies": {
-    "gatsby": "^2.17.10",
-    "gatsby-plugin-google-analytics": "^2.1.24",
+    "gatsby": "^2.20.8",
+    "gatsby-plugin-google-analytics": "^2.2.2",
     "gatsby-plugin-html-attributes": "^1.0.5",
-    "gatsby-plugin-manifest": "^2.2.26",
-    "gatsby-plugin-offline": "^3.0.18",
-    "gatsby-plugin-react-helmet": "^3.1.13",
+    "gatsby-plugin-manifest": "^2.3.3",
+    "gatsby-plugin-offline": "^3.1.2",
+    "gatsby-plugin-react-helmet": "^3.2.1",
     "gatsby-plugin-robots-txt": "^1.5.0",
-    "gatsby-plugin-sass": "^2.1.20",
-    "gatsby-plugin-sitemap": "^2.2.19",
-    "node-sass": "^4.13.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
+    "gatsby-plugin-sass": "^2.2.1",
+    "gatsby-plugin-sitemap": "^2.3.1",
+    "node-sass": "^4.13.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
     "react-helmet": "^5.2.1",
-    "react-images": "^1.0.0"
+    "react-images": "^1.1.0"
   },
   "keywords": [
     "gatsby"
@@ -26,15 +26,13 @@
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
-    "lint": "eslint ."
-  },
-  "devDependencies": {
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "prettier": "^1.19.1"
+    "format": "prettier --write \"src/components/*.js\" \"src/pages/*.js\""
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+  },
+  "devDependencies": {
+    "prettier": "2.0.2"
   }
 }
