@@ -3,7 +3,7 @@ module.exports = {
     title: "Gatsby Starter - Strata by HTML5 UP",
     author: "Hunter Chang",
     description: "A Gatsby.js Starter based on Strata by HTML5 UP",
-    siteUrl: "https://www.ghassenrjab.tn"
+    siteUrl: "https://www.ghassenrjab.tn",
   },
   plugins: [
     "gatsby-plugin-react-helmet",
@@ -16,16 +16,16 @@ module.exports = {
         background_color: "#663399",
         theme_color: "#663399",
         display: "minimal-ui",
-        icon: "src/assets/images/website-icon.png" // This path is relative to the root of the site.
-      }
+        icon: "src/assets/images/website-icon.png", // This path is relative to the root of the site.
+      },
     },
     "gatsby-plugin-sass",
     "gatsby-plugin-offline",
     {
       resolve: "gatsby-plugin-html-attributes",
       options: {
-        lang: "en"
-      }
+        lang: "en",
+      },
     },
     "gatsby-plugin-sitemap",
     "gatsby-plugin-robots-txt",
@@ -33,8 +33,8 @@ module.exports = {
       resolve: "gatsby-plugin-google-analytics",
       options: {
         trackingId: "UA-142190723-1",
-        head: true
-      }
-    }
-  ]
+        head: true,
+      },
+    },
+  ],
 };
