@@ -3,17 +3,11 @@ import "../assets/scss/main.scss";
 
 import Header from "./Header";
 
-class Template extends React.Component {
-  render() {
-    const { children } = this.props;
-
-    return (
-      <div>
-        <Header />
-        {children}
-      </div>
-    );
-  }
-}
+const Template = ({ children }) => (
+  <div>
+    <Header />
+    {children}
+  </div>
+);
 
 export default Template;
